Remove protos which are now in googleapis. (#35)
Remove protos which are now in googleapis.

Update to a more recent version of bazel for travis testing.  Update to latest versions of bazel rules for Go.  Pin to a fixed version of googleapis. Needed to update protocol buffers.  Dropping "java lite" targets as it's broken vs bazel and appears to be unsupported.