Proto support for conformance tests (#95)
Allows conformance tests to bind variables to protobuf message values, and to expect protobuf messages as evaluation results. Both proto2 and proto3 are supported.

Adds initial conformance tests for proto2 and proto3 - just enough to demonstrate that the plumbing works. More conformance tests will be added in future PRs.

The C++ filter written to parse expanded "Any" messages in textproto is no longer required and will be removed in a future PR.